+++ src/core/databasemanager.h
@@ -17,6 +17,7 @@ public:
     explicit DatabaseManager(QObject *parent = nullptr);
 
     void insertDownload(const downloadInformations &info);
+    void updateDownload(const downloadInformations &info);
     void downloadFinished(const QString &id, const QString &status);
     void removeDownload(const QString &id);
     QVector<downloadInformations> getDownloads();